Moved methods on Sensor class to SensorMixin
